refactor: refactoring form population
