+++ src/pages/Colaborador/Avaliacao.tsx
@@ -1,11 +1,8 @@
-import { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm, FormProvider } from 'react-hook-form';
 
 import { useCycle } from '../../hooks/useCycle';
-
-import type { NavigationState } from '../../hooks/useNotes';
+import { useEvaluationFormPopulation } from '../../hooks/useEvaluationFormPopulation';
 
 import { fullEvaluationSchema, type EvaluationFormData } from '../../schemas/evaluation';
 
@@ -17,8 +14,6 @@ import EvaluationSubmittedMessage from '../../components/Evaluation/EvaluationSu
 
 export function Avaliacao() {
 
-    const location = useLocation();
-
     const { currentCycle, evaluationStatus, isLoading } = useCycle();
 
     const methods = useForm<EvaluationFormData>({
@@ -26,48 +21,7 @@ export function Avaliacao() {
         mode: 'onSubmit',
     });
 
-    useEffect(() => {
-        const navigationState = location.state as NavigationState | null;
-
-        if (navigationState?.geminiResponse) {
-            const { mentoring, references, evaluation360, selfAssessment } = navigationState.geminiResponse;
-
-            if (mentoring) {
-                methods.setValue('mentoringRating', mentoring.rating ?? 0);
-                methods.setValue('mentoringJustification', mentoring.justification ?? '');
-                methods.setValue('mentoringIAValid', false, { shouldValidate: true });
-            }
-
-            if (references && references.length > 0) {
-                methods.setValue('references', references.map(ref => ({
-                    collaboratorId: ref.collaboratorId,
-                    justification: ref.justification,
-                    referencesIAValid: false
-                })));
-            }
-
-            if (evaluation360 && evaluation360.length > 0) {
-                methods.setValue('evaluation360', evaluation360.map(eval360 => ({
-                    collaboratorId: eval360.collaboratorId,
-                    rating: eval360.rating,
-                    strengths: eval360.strengths,
-                    improvements: eval360.improvements,
-                    evaluation360IAValid: false
-                })));
-            }
-
-            if (selfAssessment && selfAssessment.length > 0) {
-                methods.setValue('selfAssessment', selfAssessment.map(selfAssess => ({
-                    pilarId: selfAssess.pillarId,
-                    criterionId: selfAssess.criteriaId,
-                    rating: selfAssess.rating,
-                    justification: selfAssess.justification,
-                    selfAssessmentIAValid: false
-                })));
-            }
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [location.state]);
+    useEvaluationFormPopulation(methods);
 
     if (isLoading) {
         return <CycleLoading />;
